+++ roxctl/central/central.go
@@ -2,6 +2,8 @@ package central
 
 import (
 	"github.com/spf13/cobra"
+	"github.com/stackrox/rox/pkg/pointers"
+	"github.com/stackrox/rox/roxctl/central/backup"
 	"github.com/stackrox/rox/roxctl/central/cert"
 	"github.com/stackrox/rox/roxctl/central/db"
 	"github.com/stackrox/rox/roxctl/central/debug"
@@ -16,11 +18,11 @@ func Command() *cobra.Command {
 	c := &cobra.Command{
 		Use: "central",
 	}
-
 	c.AddCommand(
 		cert.Command(),
 		generate.Command(),
 		db.Command(),
+		backup.Command(pointers.Bool(true)),
 		debug.Command(),
 		license.Command(),
 		userpki.Command(),